move contextmenu dir
