+++ webcli/src/features/contextmenu/ContextMenuProvider.tsx
@@ -1,4 +1,4 @@
-import ContextMenu from './ContextMenu';
+import ContextMenu from './atom/ContextMenu';
 
 export default function ContextMenuProvider({ children }: { children: React.ReactNode }) {
   return (
