+++ webcli/src/features/contextmenu/atom/ContextMenuWrapper.tsx
@@ -1,6 +1,6 @@
 import Menu from '@mui/material/Menu';
 import { useAppDispatch, useAppSelector } from '~/lib/react-redux';
-import { closeContextmenu } from './contextmenuSlice';
+import { closeContextmenu } from '../contextmenuSlice';
 
 export default function ContextMenuWrapper({ children }: { children: React.ReactNode }) {
   const menuState = useAppSelector((store) => store.contextmenu.menuState);
